refactor: oauth-internal をまとめる
