+++ workspaces/client/app/routes/connect-required-oauth/page.tsx
@@ -1,7 +1,7 @@
 import {
 	OAUTH_PROVIDERS,
 	REQUIRED_OAUTH_PROVIDER_IDS,
-} from "@idp/schema/entity/oauth-provider";
+} from "@idp/schema/entity/oauth-internal/oauth-provider";
 import { useEffect } from "react";
 import { type MetaFunction, useNavigate } from "react-router";
 import { css } from "styled-system/css";
