+++ workspaces/schema/src/entity/oauth-internal/oauth-connection.ts
