+++ workspaces/schema/src/entity/oauth-internal/oauth-provider.ts
