+++ workspaces/server/src/routes/discord.ts
@@ -1,4 +1,4 @@
-import { OAUTH_PROVIDER_IDS } from "@idp/schema/entity/oauth-provider";
+import { OAUTH_PROVIDER_IDS } from "@idp/schema/entity/oauth-internal/oauth-provider";
 import {
 	OAuth2Routes,
 	OAuth2Scopes,
